moved some functions and added auth sequence annotation
